+++ tab.go
@@ -65,7 +65,7 @@ func (t *Tab) InsertAt(name string, n int, child Control) {
 // Delete deletes the nth page of the Tab.
 func (t *Tab) Delete(n int) {
 	t.children = append(t.children[:n], t.children[n + 1:]...)
-	C.uiTabDelete(t.t, C.uintmax_t(n))
+	C.uiTabDelete(t.t, C.int(n))
 }
 
 // NumPages returns the number of pages in the Tab.
@@ -76,12 +76,12 @@ func (t *Tab) NumPages() int {
 // Margined returns whether page n (starting at 0) of the Tab
 // has margins around its child.
 func (t *Tab) Margined(n int) bool {
-	return tobool(C.uiTabMargined(t.t, C.uintmax_t(n)))
+	return tobool(C.uiTabMargined(t.t, C.int(n)))
 }
 
 // SetMargined controls whether page n (starting at 0) of the Tab
 // has margins around its child. The size of the margins are
 // determined by the OS and its best practices.
 func (t *Tab) SetMargined(n int, margined bool) {
-	C.uiTabSetMargined(t.t, C.uintmax_t(n), frombool(margined))
+	C.uiTabSetMargined(t.t, C.int(n), frombool(margined))
 }
